+++ components/RegisterSystemForm/RegisterEndpoints/hook.ts
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { SystemService } from "@services/system";
 import { useRouter } from "next/router";
 import { useSystemRegistrationContext } from "@contexts/SystemRegistrationContext";
-import { isApiError } from "@common/api";
+import { ApiError, isApiError } from "@common/api";
 
 export const useRegisterEndpoints = () => {
   const router = useRouter();
@@ -38,7 +38,7 @@ export const useRegisterEndpoints = () => {
     setLoading(false);
 
     if (isApiError(response)) {
-      showMessage(response.error);
+      showMessage((response as ApiError).error);
     } else {
       Object.keys(serviceToOpenApiFilename).forEach((service: string) => {
         setServiceToSynAndAsyncOperations(service, {
